Now the application checks wether registering should be enabled or disabled depending on an environment variable that marks the opening date.
A new middleware function is responsible for checking if registrations
must be open or not and add the info to the rendering options for that
request, together to the user data (if the user is logged in)

Rendering options are now added to an object in req.app.locals, so
that the new middleware that was introduced can access it.

The CSS for the app has been completely revamped, so that a dropdown
menu appears when the device in which the app is displaying is too
small to fit the full appearance.